% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monotone_bart.R
\name{monotone_bart}
\alias{monotone_bart}
\title{BART for binary outcomes with predicted probabilities monotone in one binary covariate}
\usage{
monotone_bart(y, z, x, xpred, nskip = 5000, ndpost = 5000, m = 50)
}
\arguments{
\item{y}{The response variable (binary)}

\item{z}{The monotone covariate (binary)}

\item{x}{The design matrix}

\item{xpred}{The design matrix for predictions}

\item{nskip}{MCMC burn-in period}

\item{ndpost}{MCMC samples to return}

\item{m}{Number of trees}
}
\value{
A list with elements pr0 and pr1, which are samples of P(Y=1|X=x, Z=z) for z=0 and 1 for x in xpred
}
\description{
BART for binary outcomes with predicted probabilities monotone in one binary covariate
}
